+++ backend.js
@@ -6,23 +6,9 @@ var zlib = require('zlib');
 
 module.exports = Backend;
 
-function Task(callback) {
-    this.err = null;
-    this.headers = {};
-    this.access = +new Date;
-    this.done;
-    this.body;
-    this.once('done', callback);
-};
-util.inherits(Task, require('events').EventEmitter);
-
 function Backend(opts, callback) {
-    this._vectorCache = {};
-    this._vectorTimeout = null;
     this._scale = opts.scale || 1;
-    this._maxAge = typeof opts.maxAge === 'number' ? opts.maxAge : 60e3;
     this._deflate = typeof opts.deflate === 'boolean' ? opts.deflate : true;
-    this._reap = typeof opts.reap === 'number' ? opts.reap : 60e3;
     this._source = null;
     var backend = this;
 
@@ -83,31 +69,6 @@ Backend.prototype.getTile = function(z, x, y, callback) {
         by = Math.floor(y / Math.pow(2, z - bz));
     }
 
-    var key = bz + '/' + bx + '/' + by;
-    var cache = backend._vectorCache[key];
-
-    // Reap cached vector tiles with stale access times on an interval.
-    if (backend._reap && !backend._vectorTimeout) backend._vectorTimeout = setTimeout(function() {
-        var now = +new Date;
-        Object.keys(backend._vectorCache).forEach(function(key) {
-            if ((now - backend._vectorCache[key].access) < backend._maxAge) return;
-            delete backend._vectorCache[key];
-        });
-        delete backend._vectorTimeout;
-    }, backend._reap);
-
-    // Expire cached tiles when they are past maxAge.
-    if (cache && (now-cache.access) >= backend._maxAge) cache = false;
-
-    // Return cache if finished.
-    if (cache && cache.done) return callback(null, cache.body, cache.headers);
-
-    // Otherwise add listener if task is in progress.
-    if (cache) return cache.once('done', callback);
-
-    var task = new Task(callback);
-    backend._vectorCache[key] = task;
-
     var size = 0;
     var headers = {};
 
@@ -120,7 +81,7 @@ Backend.prototype.getTile = function(z, x, y, callback) {
             by = Math.floor(y / Math.pow(2, z - bz));
             return source.getTile(bz, bx, by, sourceGet);
         }
-        if (err && err.message !== 'Tile does not exist') return done(err);
+        if (err && err.message !== 'Tile does not exist') return callback(err);
 
         if (!body) {
             return makevtile();
@@ -135,16 +96,8 @@ Backend.prototype.getTile = function(z, x, y, callback) {
         }
     });
 
-    function done(err, body, headers) {
-        if (err) delete backend._vectorCache[key];
-        task.done = true;
-        task.body = body;
-        task.headers = headers;
-        task.emit('done', err, body, headers);
-    };
-
     function makevtile(err, data) {
-        if (err && err.message !== 'Tile does not exist') return done(err);
+        if (err && err.message !== 'Tile does not exist') return callback(err);
 
         // If no last modified is provided, use epoch.
         headers['Last-Modified'] = new Date(headers['Last-Modified'] || 0).toUTCString();
@@ -158,22 +111,22 @@ Backend.prototype.getTile = function(z, x, y, callback) {
         headers['Content-Type'] = 'application/x-protobuf';
 
         // Pass-thru of an upstream mapnik vector tile (not pbf) source.
-        if (data instanceof mapnik.VectorTile) return done(null, data, headers);
+        if (data instanceof mapnik.VectorTile) return callback(null, data, headers);
 
         var vtile = new mapnik.VectorTile(bz, bx, by);
         vtile._srcbytes = size;
 
         // null/zero length data is a solid tile be painted.
-        if (!data) return done(null, vtile, headers);
+        if (!data || !data.length) return callback(null, vtile, headers);
 
         try {
             vtile.setData(data);
         } catch (err) {
-            return done(err);
+            return callback(err);
         }
         vtile.parse(function(err) {
-            if (err) return done(err);
-            done(null, vtile, headers);
+            if (err) return callback(err);
+            callback(null, vtile, headers);
         })
     };
 };